requirements.txt
faiss_cpu==1.10.0
numpy==2.2.6
scikit_learn==1.6.1
sentence_transformers==4.53.0
tabulate==0.9.0
